# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for income and expense applications that allows you to receive a report for a certain period",
        "title": "Expense Application",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080/api",
    "basePath": "/",
    "paths": {
...
        "/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that return list of roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "get-roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RolesResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that store role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "store-roles",
                "parameters": [
                    {
                        "description": "Role form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    }
                }
            }
        },
        "/roles/permissions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that return list of permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "get-permissions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PermissionsResponse"
                        }
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that return role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "get-role",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RoleResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that allow to update role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "update-roles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that allow to delete role by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "delete-roles",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Role ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/users/:id/assign-role": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": [
                            "admin"
                        ]
                    }
                ],
                "description": "Method that return list of permissions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "operationId": "assign-role-to-users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Role form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/AssignRoleToUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    }
                }
            }
        },
        "/users/current": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method allow to get current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "operationId": "users-current",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/UserResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Method allow to update users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "operationId": "users-update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Users ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Status"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "AssignRoleToUserRequest": {
            "type": "object",
            "properties": {
                "role_id": {
                    "type": "integer",
                    "example": 14
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Something went wrong try again later!"
                }
            }
        },
        "PermissionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "prefix_create",
                        "prefix_read",
                        "..."
                    ]
                }
            }
        },
        "Role": {
            "type": "object",
            "properties": {
                "display_title": {
                    "type": "string",
                    "example": "UserTitle"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "prefix_create",
                        "prefix_read"
                    ]
                }
            }
        },
        "RoleResponse": {
            "type": "object",
            "properties": {
                "display_title": {
                    "type": "string",
                    "example": "UserTitle"
                },
                "id": {
                    "type": "integer",
                    "example": 12
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/PermissionsResponse"
                    }
                },
                "title": {
                    "type": "string",
                    "example": "RoleTitle"
                }
            }
        },
        "RolesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RoleResponse"
                    }
                }
            }
        },
...
        "Status": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "Everything is OK"
                }
            }
        },
        "UserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jon_doe@gmail.com"
                },
                "id": {
                    "type": "integer",
                    "example": 14
                },
                "name": {
                    "type": "string",
                    "example": "Jon Doe"
                },
                "role": {
                    "$ref": "#/definitions/RoleResponse"
                },
                "tg_id": {
                    "type": "integer",
                    "example": 131231231231323
                }
            }
        },
        "UserUpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "jon_doe@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "Jon Doe"
                },
                "password": {
                    "type": "string",
                    "example": "qwerty"
                },
                "tg_id": {
                    "type": "integer",
                    "example": 131231231231323
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}